AI_picture_creater/requirements.txt
flask==2.3.3
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0